01-CoreConcepts/requirements.txt
fastmcp>=1.0.0
python-dotenv>=1.1.0
openai>=1.12.0